Finished fixing trim_nan_edges and updated code cell with function.
There was a math error. We were using slicing to try to find the first index where leading NaNs stopped and the last index before lagging NaNs started. But using slicing caused the code to create a shortened list in the edge case where there are no lagging NaNs. So to deal with this issue for lagging NaNs, I switched to masking to find the true values and using .index[-1] to find the first non NaN going from the bottom to the top of the list.