Just found out my postinjury tall data sheet is inaccurate.
All animal data is in the combined data sheet that gets reformatted into the tall data sheet. For whatever reason, it seems that the data in the combined datasheet has to be in a rectangular shape. There can't be one cohort with more data than the others. That's why there are leading and lagging NaNs at the end of every animal's data. I went back to the combined datasheet to add cohort 8 and re-run the macro so that I could complete data cleaning. When doing so, I noticed that cohort 7 is longer than all the other cohorts, which means that the data does not have a rectangular shape. This throws off the macro such that the macro truncates data for the longest cohort to the length of all the other cohorts.  I need to look at the macro to determine whether there is a line in the macro that specifies this behavior so that I can either change the datasheet or the macro.