Worked towards k-means clustering. Still in preprocessing stage. Have to convert data from TIDY format to WIDE format.
Saved postinjury files. Had a really good idea, which was to output all the text that I write to myself to remind myself what the output of a code cell means as its own file, as a kind of report for later. Will definitely do that for the next file but may return to create for this file as well.